requirements.txt
bleach==2.1.3
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
crcmod==1.7
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.10.8
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
django-formtools==2.0
django-guardian==1.4.9
django-haystack==2.8.1
django-import-export==1.0.0
django-qiniu-storage==2.3.1
django-redis==4.8.0
django-reversion==2.0.9
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-haystack==1.7.1rc2
et-xmlfile==1.0.1
future==0.16.0
gunicorn==19.7.1
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
oauthlib==2.0.6
odfpy==1.3.6
olefile==0.44
openpyxl==2.5.0
Pillow==4.2.1
pycryptodome==3.4.6
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
qiniu==7.2.0
raven==6.2.1
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.1.0
XlsxWriter==0.9.8
xlwt==1.2.0
